Move return statements that prevent filling the Ntuples in VHbbAnalyzer
If passall is enabled, the 4 lines sometimes caused too early exit in
the analysis chain and some of the Ntuple branches wouldn't be filled
correctly because of it. The solution is to check first whether all events
are supposed to pass the cuts or not (which is the default behavior).